fix(ac-626): align capability provider inventory
